Rookie Created Gestures
C# project that rookies started making gestures in last night.
Contains the latest build of the Kinect Gesture Base Lib thing so that
they can easily extend off of my Abstract Gesture class and not have to
worry about NUI, refreshing, etc.
They wrote a basic Arm Above head which works fine, and they are in the
process of debugging a gesture that is used on the Xbox.
They're all pretty self-sufficient at this point which is also good :)